refactor(SimulationEntities): move classes
