+++ src/main/scala/ulisse/entities/route/RouteEnvironmentElement.scala
@@ -1,0 +1,7 @@
+package ulisse.entities.route
+
+import ulisse.entities.route.Routes.Route
+import ulisse.entities.train.Trains.Train
+
+trait RouteEnvironmentElement extends Route:
+  val trains: Seq[Seq[Train]]
