+++ src/main/scala/ulisse/entities/simulation/Agents.scala
@@ -7,8 +7,3 @@ import ulisse.entities.train.Trains.Train
 object Agents:
   trait SimulationAgent:
     def doStep(dt: Int, simulationEnvironment: SimulationEnvironment): SimulationAction
-
-  trait TrainAgent extends Train with SimulationAgent:
-    def travelDistance: Double
-    def travelDistance_=(newTravelDistance: Double): TrainAgent
-    def updateTravelDistance(distanceDelta: Double): TrainAgent
